+++ src/app.module.css
@@ -1,71 +1,67 @@
-@import url('https://fonts.googleapis.com/css2?family=Pixelify+Sans&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Pixelify+Sans&display=swap");
 
-body{
-    background-image: url('./assets/Diseño sin título.gif');
-    background-size: 'cover';
-      height: '100vh';
-      width: '100vw';
-      font-family: 'Pixelify Sans', sans-serif;
+body {
+  background-image: url("./assets/Diseño sin título.gif");
+  background-size: "cover";
+  height: "100vh";
+  width: "100vw";
+  font-family: "Pixelify Sans", sans-serif;
 }
 
 .container {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: column;
-    justify-content: flex-start;
-    
+  display: flex;
+  height: 93.5vh;
+  overflow-x: hidden;
+  flex-wrap: wrap;
+  flex-direction: column;
+  justify-content: flex-start;
 }
 
 .container img {
-    width: 50%;
-    margin: 35px 0;
+  width: 50%;
+  margin: 35px 0;
 }
 
 .containerMini {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
+  display: flex;
+  flex-wrap: wrap;
 }
 
 .containerMini div {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 
 .containerMini div p {
-    margin-top: 0px; 
+  margin-top: 0px;
 }
 
 /* Rocket animation */
 .rocket {
-    position: fixed;
-    bottom: 0;
-    top:45%;
-    left: 50%;
-    transform: translate(-50%, 100%);
-    animation: rocket 3s ease-in-out forwards;
-  }
-  
-  @keyframes rocket {
-    0% {
-      opacity: 0;
-      transform: translate(-50%, 100%);
-      width: 450px;
-      height: 600px;
-    }
-    100% {
-        opacity: 1;
-        transform: translate(-50%, -50%);
-        width: 450px;
-        height: 800px;
-    }
-  }
+  position: fixed;
+  bottom: 0;
+  top: 45%;
+  left: 50%;
+  transform: translate(-50%, 100%);
+  animation: rocket 3s ease-in-out forwards;
+}
 
-  .Nav{
-    width: '100vw';
+@keyframes rocket {
+  0% {
+    opacity: 0;
+    transform: translate(-50%, 100%);
+    width: 450px;
+    height: 600px;
   }
-  /* Fire animation */
+  100% {
+    opacity: 1;
+    transform: translate(-50%, -50%);
+    width: 450px;
+    height: 800px;
+  }
+}
+/* Fire animation */
 .fire {
   position: fixed;
   /*top: 100%; Cambia 'top' por 'bottom'*/
