+++ src/componentes/footerNav/NavFooter.module.css
@@ -1,3 +1,28 @@
-.container{
+.footer{
     background-color: gray;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    padding: 4px 15px 0px 15px;
+    color: black;
+}
+.start{
+    position: relative;
+    margin-top: -34px;
+    width: 80px;
+    height: 24px;
+    padding-right: 4px ;
+    text-align: right;
+    border-right: 2px solid black;
+    border-bottom: 2px solid black;
+    border-left: 2px solid rgb(188, 185, 185);
+    border-top: 2px solid rgb(188, 185, 185);
+}
+.win{
+    position: relative;
+    padding-top: 30px;
+}
+.clock{
+    margin-top: -1px;
+
 }
